+++ backend/infrastructures/extract/strategies.py
@@ -1,0 +1,22 @@
+import modules.extract.extract  as ex
+from pathlib import Path
+import logging
+import pandas as pd
+
+logger = logging.getLogger(_name_)
+
+class ExtractCsvStrategy(ex.ExtractStrategy):
+    def extract(path_to_file: Path):
+        logger.info("loading csv...")
+        return pd.read_csv(path_to_file)
+
+
+class ExtractExcelStrategy(ex.ExtractStrategy):
+    def extract(path_to_file: Path):
+        logger.info("loading excel...")  
+        return pd.read_excel(path_to_file)
+
+
+class ExtractPdfStrategy(ex.ExtractStrategy):
+    def extract(path_to_file: Path):
+        raise NotImplementedError("PDF not implemented yet")
