+++ backend/modules/transform/transform.py
@@ -17,41 +17,3 @@ class EtlHandler(ABC):
 
     def next(self, data: pd.DataFrame) -> pd.DataFrame:
         return self.next_handler.process() if self.next_handler != None else data
-
-
-class ExtractHandler(EtlHandler):
-    def __init__(self, path_to_file: str, next_handler: EtlHandler = None):
-        self.path_to_file = path_to_file
-        self.next_handler = next_handler
-
-    def process(self, data=None):
-        data_file_path = Path(self.path_to_file)
-        file_extension = data_file_path.suffix()
-        strat = self._chooseStrategy(file_extension)
-        extract_context = extract.ExtractContext(strat)
-        df = extract_context.extract(self.path_to_file)
-        return self.next(df)
-
-    def _chooseStrategy(file_extension: str) -> extract.ExtractStrategy:
-        if file_extension.lower() == ".csv":
-            return extract.ExtractCsvStrategy()
-        elif file_extension.lower() in [".xls", ".xlsx"]:
-            return extract.ExtractExcelStrategy()
-        elif file_extension.lower() in [".pdf"]:
-            return extract.ExtractPdfStrategy
-        else:
-            raise TypeError(file_extension)
-
-
-class ProcessHandler(EtlHandler):
-    def __init__(self, next_handler: EtlHandler = None):
-        self.next_handler = next_handler
-
-    def process(self, df: pd.DataFrame):
-        if df == None:
-            raise ValueError("Input DataFrame doesn't exist")
-        # process handler logic to be discussed.
-        return self.next(df)
-
-
-# TODO - add Persistance Handler
